+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Development version
+New methods:
+
+- Added ``delaunay_triangles`` method from shapely to GeoSeries/GeoDataframe.
 
 New features and improvements: